Separate Configuration Channel handler
